Introduce base class to simplify extra rules
